using change password pop up component + change to authorize
