+++ vegefoods.angularui/src/app/Pages/partials/changepassword/changepassword.component.html
@@ -1,32 +1,41 @@
 <!-- Change Password Pop up Start -->
-<div class="container-fluid page-padding-top">
-    <h1 class="ms-5 ps-5 text-secondary">Change Password</h1>  
-    <div class="container">
-        <form [formGroup]="changePasswordForm" >
-            <input type="hidden" value="{{userId}}" className="form-control" />
-            <div class="row g-5">
-                <div class="col-md-12 col-lg-6 col-xl-7">
-                    <div class="form-item">
-                        <label class="form-label my-3">Current Password<sup>*</sup></label>
-                        <input type="email"  class="form-control">
-                    </div>
-                    <div class="form-item">
-                        <label class="form-label my-3">New Password<sup>*</sup></label>
-                        <input type="email"  class="form-control">
-                    </div>
-                    <div class="form-item">
-                        <label class="form-label my-3">New Password<sup>*</sup></label>
-                        <input type="password"  class="form-control">
-                    </div>
-                            
-                    <div class="mt-4">                        
-                        <button class="btn border-secondary rounded-pill px-4 py-3 text-primary" type="submit">Save</button>
-                        &nbsp;&nbsp;&nbsp;
-                        <button class="btn border-secondary rounded-pill px-4 py-3 text-danger" type="button">Cancel</button>
-                    </div>       
-                </div>                        
-            </div>
-        </form>
+<div class="modal-dialog">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h4 class="modal-title">Change Password</h4>
+        <button type="button" class="btn-close" (click)="CloseModel()" data-bs-dismiss="modal"></button>
+      </div>
+      <div class="modal-body">
+          <div class="container-fluid">
+              <div class="container">
+                  <form [formGroup]="userPasswordForm"  (ngSubmit)="changePassword()">
+                      <input type="hidden" formControlName="id" className="form-control" />
+                      <div class="row g-5">
+                          <div class="col-md-12 col-lg-6 col-xl-7">
+                              <div class="form-item">
+                                  <label class="form-label my-3">Current Password<sup>*</sup></label>
+                                  <input type="password" formControlName="currentpassword" class="form-control">
+                              </div>
+                              <div class="form-item">
+                                  <label class="form-label my-3">New Password<sup>*</sup></label>
+                                  <input type="password" formControlName="newpassword" class="form-control">
+                              </div>
+                              <div class="form-item">
+                                  <label class="form-label my-3">New Password<sup>*</sup></label>
+                                  <input type="password" formControlName="confirmpassword" class="form-control">
+                              </div>
+                                      
+                              <div class="mt-4">                        
+                                  <button class="btn border-secondary rounded-pill px-4 py-3 text-primary" type="submit">Save</button>
+                                  &nbsp;&nbsp;&nbsp;
+                                  <button class="btn border-secondary rounded-pill px-4 py-3 text-danger" data-bs-dismiss="modal" type="button">Cancel</button>
+                              </div>       
+                          </div>                        
+                      </div>
+                  </form>
+              </div>
+          </div>
+      </div>
     </div>
-</div>
+  </div>
 <!-- Change Password Pop up End -->
